Working on license verification.

+++ launcher/src/launcher/Updater.java
@@ -1,6 +1,7 @@
 package launcher;
 
 import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.IOUtils;
 import sps.core.Logger;
 
 import java.io.File;
@@ -12,12 +13,15 @@ public class Updater {
     private File updateDir = new File("aigilas-update");
     private String licenseText;
 
-    public Updater(){
+    public Updater() {
 
     }
 
     public boolean runIfNeeded(String license) {
-        boolean success = checkLicense(license) && checkVersion() && downloadUpdate() && applyUpdate();
+        boolean success = false;
+        if(checkLicense(license) && checkVersion() && downloadUpdate() && applyUpdate()){
+            success = true;
+        }
         clean();
         Logger.info("Finished applying updates.");
         return success;
@@ -25,11 +29,17 @@ public class Updater {
 
 
     private boolean checkLicense(String license) {
-        Logger.info("Validating SE license");
-        Logger.info("Checking to see if a stable edition license has been entered.");
-        if (licenseText != null && !licenseText.isEmpty()) {
-            //TODO Validate against the server
-            return true;
+        try {
+            Logger.info("Validating SE license");
+            Logger.info("Checking to see if a stable edition license has been entered.");
+            if (licenseText != null && !licenseText.isEmpty()) {
+                URL licenseCheckUrl = new URL("http://www.simplepathstudios.com");
+                String response = IOUtils.toString(licenseCheckUrl.openStream());
+                return response.contains("true");
+            }
+        }
+        catch (Exception e) {
+            Logger.exception(e);
         }
         return false;
     }
@@ -45,7 +55,9 @@ public class Updater {
 
             URL versionCheckUrl = new URL("http://www.simplepathstudios.com/download.php?target=aigilas&version=" + myVersion);
             //TODO This is most likely not the actual way to query
-            String result = versionCheckUrl.getContent().toString();
+
+            String result = IOUtils.toString(versionCheckUrl.openStream());
+            Logger.info("Result is: " + result);
             return result.contains("true");
         }
         catch (Exception e) {
@@ -54,7 +66,7 @@ public class Updater {
         return true;
     }
 
-    private boolean downloadUpdate(){
+    private boolean downloadUpdate() {
         try {
             Logger.info("Preparing the update location");
 
@@ -97,8 +109,12 @@ public class Updater {
     private boolean clean() {
         try {
             Logger.info("Cleaning up temporary files");
-            FileUtils.forceDelete(update);
-            FileUtils.deleteDirectory(updateDir);
+            if (update.exists()) {
+                FileUtils.forceDelete(update);
+            }
+            if (updateDir.exists()) {
+                FileUtils.deleteDirectory(updateDir);
+            }
         }
         catch (Exception e) {
             Logger.exception(e);
